+++ src/wurf/http_resolver.py
@@ -10,19 +10,16 @@ class HttpResolver(object):
     Http Resolver functionality. Downloads a file.
     """
 
-    def __init__(self, url_download, dependency, source, cwd):
-
+    def __init__(self, url_download, dependency, cwd):
         """Construct a new instance.
 
         :param url_download: An UrlDownload instance
         :param dependency: The dependency instance.
-        :param source: The URL of the dependency as a string
         :param cwd: Current working directory as a string. This is the place
             where we should create new folders etc.
         """
         self.url_download = url_download
         self.dependency = dependency
-        self.source = source
         self.cwd = cwd
 
     def resolve(self):
@@ -31,12 +28,10 @@ class HttpResolver(object):
 
         :return: The path to the resolved dependency as a string.
         """
-        # Store the current source in the dependency object
-        self.dependency.current_source = self.source
-
         # Use the first 6 characters of the SHA1 hash of the repository url
         # to uniquely identify the repository
-        source_hash = hashlib.sha1(self.source.encode("utf-8")).hexdigest()[:6]
+        source = self.dependency.source
+        source_hash = hashlib.sha1(source.encode("utf-8")).hexdigest()[:6]
 
         # The folder for storing the file
         folder_name = "http-" + source_hash
@@ -51,7 +46,7 @@ class HttpResolver(object):
             filename = None
 
         file_path = self.url_download.download(
-            cwd=folder_path, source=self.source, filename=filename
+            cwd=folder_path, source=source, filename=filename
         )
 
         assert os.path.isfile(file_path), "We should have a valid path here!"