+++ test/python/test_http_resolver.py
@@ -5,16 +5,14 @@ from wurf.http_resolver import HttpResolver
 
 
 def test_http_resolver(testdirectory):
-
     url_download = mock.Mock()
     dependency = mock.Mock()
     dependency.filename = None
-
-    http_source = "http://example.com/file.zip"
+    dependency.source = "http://example.com/file.zip"
     cwd = testdirectory.path()
 
     def create_file(cwd, source, filename):
-        assert http_source == source
+        assert dependency.source == source
         assert filename is None
 
         httpdir = testdirectory.from_path(cwd)
@@ -24,9 +22,7 @@ def test_http_resolver(testdirectory):
 
     url_download.download.side_effect = create_file
 
-    resolver = HttpResolver(
-        url_download=url_download, dependency=dependency, source=http_source, cwd=cwd
-    )
+    resolver = HttpResolver(url_download=url_download, dependency=dependency, cwd=cwd)
 
     path = resolver.resolve()
     assert os.path.isfile(path)
@@ -35,16 +31,15 @@ def test_http_resolver(testdirectory):
 
 
 def test_http_resolver_filename(testdirectory):
-
     url_download = mock.Mock()
     dependency = mock.Mock()
     dependency.filename = "foo.zip"
 
-    http_source = "http://example.com/file.zip"
+    dependency.source = "http://example.com/file.zip"
     cwd = testdirectory.path()
 
     def create_file(cwd, source, filename):
-        assert http_source == source
+        assert dependency.source == source
         assert filename == "foo.zip"
 
         httpdir = testdirectory.from_path(cwd)
@@ -54,9 +49,7 @@ def test_http_resolver_filename(testdirectory):
 
     url_download.download.side_effect = create_file
 
-    resolver = HttpResolver(
-        url_download=url_download, dependency=dependency, source=http_source, cwd=cwd
-    )
+    resolver = HttpResolver(url_download=url_download, dependency=dependency, cwd=cwd)
 
     path = resolver.resolve()
     assert os.path.isfile(path)